Move toolbar to application_widgets
